Refactor desktop chat shared UI and streaming state
- share the chat bubble, typing indicator, notch thinking mark, streaming buffer, tool block updater, and follow-up queue across main chat and task chat
- preserve main-chat-only behavior through normalizer hooks, generated-resource attachment, metadata, ratings, citations, and stall UI
- handle task-chat failure/stop edge cases so partial text remains visible and intentional stops do not surface failure text

Oracle review: follow-up review found no structural blocker; applied recommended fixes for forced-flush normalization, task stop failure text, partial text visibility, and task follow-up scroll token timing.

Verification:
- cd desktop/macos && xcrun swift build -c debug --package-path Desktop
- cd desktop/macos && xcrun swift test -c debug --package-path Desktop --filter DesktopChatDriftGuardTests
- cd desktop/macos && xcrun swift test -c debug --package-path Desktop --filter AgentPillLifecycleTests/testTypingIndicatorUsesSharedNotchThinkingMarkInsteadOfBounce
- cd desktop/macos && xcrun swift test -c debug --package-path Desktop --filter ToolCallStatusTests
- cd desktop/macos && xcrun swift test -c debug --package-path Desktop --filter TaskChatKernelIdentityTests
- git diff --check
