+++ desktop/macos/changelog/unreleased/20260707-chat-typing-indicator.json
@@ -1,0 +1,3 @@
+{
+  "change": "Improved the desktop chat loading indicator so it matches the calmer notch thinking animation"
+}
